Move test_endpoint_get_current_build to its own file
